+++ src/Http/Controllers/Admin/PageController.php
@@ -19,7 +19,7 @@ class PageController extends BaseAdminController
 {
     use BlogPageCommon;
 
-    protected $dataFields = ['id','lang.title as title','lang.name as name', 'category.link_rewrite as category', 'link_rewrite', 'publish_status'];
+    protected $dataFields = ['id','lang.title as title','lang.name as name', 'category.link_rewrite as category', 'link_rewrite', 'read_count', 'publish_status', 'updated_at'];
 
     protected $dataSource = Page::class;